# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API Documentation",
        "description": "Dokumentasi API untuk aplikasi E-Ticketing Helpdesk berbasis Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menampilkan daftar notifikasi",
                "description": "Mengambil daftar notifikasi milik user yang sedang login.",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "responses": {
                    "200": {
                        "description": "Daftar notifikasi berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menghapus semua notifikasi",
                "description": "Menghapus semua notifikasi milik user yang sedang login.",
                "operationId": "2f7f8f22ee8678ec5d615a0904dea350",
                "responses": {
                    "200": {
                        "description": "Semua notifikasi berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menandai satu notifikasi sudah dibaca",
                "description": "Mengubah status satu notifikasi menjadi sudah dibaca berdasarkan ID.",
                "operationId": "ca686e58b97d8e0128abdaaa5a7d533a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notifikasi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifikasi berhasil ditandai sudah dibaca"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Notifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menandai semua notifikasi sudah dibaca",
                "description": "Mengubah semua notifikasi user menjadi sudah dibaca.",
                "operationId": "b9993ef4e144277796573c1b94023ac4",
                "responses": {
                    "200": {
                        "description": "Semua notifikasi berhasil ditandai sudah dibaca"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menghapus satu notifikasi",
                "description": "Menghapus notifikasi berdasarkan ID.",
                "operationId": "ab7264b840570ab1351fff8ba6732bdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notifikasi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifikasi berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Notifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Menampilkan profil user",
                "description": "Mengambil data profil user yang sedang login.",
                "operationId": "f1784685f8270527da017b3616288443",
                "responses": {
                    "200": {
                        "description": "Profil berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profil user",
                "description": "Memperbarui data profil user yang sedang login.",
                "operationId": "875e8801180287e4193f1aced9023960",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Annisa Putri"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "annisa@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/change-password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mengubah password",
                "description": "Mengubah password user yang sedang login.",
                "operationId": "a538d8818d17f49848831704e0aa7531",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "oldpassword123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diubah"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/helpdesks": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Menampilkan daftar helpdesk",
                "description": "Mengambil daftar user dengan role helpdesk untuk kebutuhan assignment tiket.",
                "operationId": "7907597d49ae3c91ad9fda7ad92cc5ec",
                "responses": {
                    "200": {
                        "description": "Daftar helpdesk berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}